% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Accessible Shiny App}
\usage{
app(title = "", lang = "en", direction = "ltr", classnames = NULL, ...)
}
\arguments{
\item{title}{a title for the document}

\item{lang}{specify the language of the document}

\item{direction}{specify the direction in which the document should be read
Use "rtl" (right to left), "ltr" (left to right), or "auto"}

\item{classnames}{css classes to apply to the parent container}

\item{...}{the content of the shiny app}
}
\value{
load package assets into the app and define the document attributes
}
\description{
Primary function for initiating the accessible shiny package client side.
This function is required in order to use this package. You can also use
this function to document attributes, such as language.
}
\examples{
if (interactive()) {
  ui <- accessibleshiny::app(
    title = "My Shiny App",
    lang = "en",
    direction = "ltr",
    tags$h1("Hello, world!")
  )

  server <- function(input, output, session) {
  }

  shinyApp(ui, server)
}

}
\references{
\url{https://developer.mozilla.org/en-US/docs/Web/API/Document/title}
\url{https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/dir}
\url{https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/lang}
}
