% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_input.R
\name{clear_accordion_input}
\alias{clear_accordion_input}
\title{\code{clear_accordion_input}}
\usage{
clear_accordion_input(inputId)
}
\arguments{
\item{inputId}{the ID of the component to clear}
}
\description{
A server-side function that clears the accordion input component
regardless of the default setting. Running this function will collapse
the accordion and uncheck the input.
}
