% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{datatable}
\alias{datatable}
\title{Accessible and Responsive Datatables}
\usage{
datatable(
  data,
  id = NULL,
  caption = NULL,
  css = NULL,
  options = list(responsive = TRUE, rowHeaders = TRUE, asHTML = FALSE)
)
}
\arguments{
\item{data}{A data object used to render the table (required)}

\item{id}{A unique id for the table (required)}

\item{caption}{A short description (1-2 lines) for the table (optional)}

\item{css}{A character string of css name(s) to render in the table element}

\item{options}{A list of options to pass on to the render table method
@param responsive A logical arg for turning on/off the rendering of
       additional elements for a responsive tables (default = FALSE)
@param rowHeaders a logical argument when true sets the first cell
       of every row as a header element (default = FALSE).
@asHTML A logical argument when true will render cell values as
        html elements (default = FALSE)}
}
\value{
Returns an html object, i.e., shiny tagList. Use
        \code{options = list(...)} for addtional rendering options.
}
\description{
The \code{datatable} function creates an accessible, responsive table from a
dataset. The function returns a shiny tagList object which can
be used in shiny applications or markdown documents. This function
can also be used as an html table generator and the output can be
written to file. This function takes the following arguments.
}
\section{Arugments}{
NA
}

\examples{
datatable(data = iris, id = "iris-table")

datatable(data = iris, id = "iris-table", css = "dark-theme")

datatable(id = "iris", data = iris, options = list(responsive = T))

datatable(id = "iris", data = iris, options = list(asHTML = T))

tbl <- datatable(id = "iris", data = iris, options = list(asHTML = T))
writeLines(as.character(tbl), "~/Desktop/iris_table.html")

}
\author{
dcruvolo
}
\keyword{a11y}
\keyword{datatable}
