% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icon$checkmark}
\alias{icon$checkmark}
\title{\code{checkmark}}
\usage{
icon$checkmark()
}
\arguments{
\item{id}{a unique ID to be applied to the svg element}

\item{class}{a css class to be applied to the svg element}

\item{background_fill}{a color to be applied to the inner svg element}

\item{icon_color}{a color to be applied to the inner svg element}

\item{aria_hidden}{a logical value for hiding elements
(Useful for purely aesthetic elements)}
}
\value{
Create an svg icon of a checkmark
}
\description{
\code{checkmark}
}
\keyword{checkmark}
\keyword{icons}
