% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icon$chevron}
\alias{icon$chevron}
\title{\code{chevron}}
\usage{
icon$chevron
}
\arguments{
\item{id}{a unique ID to be applied to the svg element}

\item{class}{a css class to be applied to the svg element}

\item{icon_fill}{a color that fills the icon shape
(useful when closed = TRUE)}

\item{icon_color}{a color to be applied to the inner svg element}

\item{closed}{a logical arg that created a closed shape (default: false)}

\item{direction}{a value that specifies the direction the chevron points to
("bottom", "top", "left", "right")}

\item{data_group}{a custom data attribute}

\item{aria_hidden}{a logical value for hiding elements
(Useful for purely aesthetic elements)}
}
\value{
An svg chevron
}
\description{
\code{chevron}
}
\keyword{chevron}
\keyword{icons}
