% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icons$plus}
\alias{icons$plus}
\title{\code{plus}
A plus sign with a filled background}
\usage{
icons$plus()
}
\arguments{
\item{id}{a unique ID for the icon}

\item{class}{a css classname to assign to the icon}

\item{background_fill}{a background color for the icon (default: #f6f6f6)}

\item{icon_color}{a color for the plus symbol (default: #3f454b)}

\item{data_group}{a custom data attribute for use in the accordion function}
}
\value{
A plus sign with a filled background
}
\description{
\code{plus}
A plus sign with a filled background
}
\keyword{icons,}
\keyword{plus}
