% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icons$restart}
\alias{icons$restart}
\title{\code{restart}}
\usage{
icons$restart()
}
\arguments{
\item{id}{a unique ID to be applied to the svg element}

\item{class}{a css class to be applied to the svg element}

\item{icon_color}{a color to be applied to the inner svg element}

\item{data_group}{a custom data attribute useful for subgroups of icons}

\item{aria_hidden}{a logical value for hiding elements
(Useful for purely aesthetic elements)}
}
\value{
create a restart icon
}
\description{
\code{restart}
}
\keyword{icons}
\keyword{resfresh}
\keyword{restart}
