% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icons$warning}
\alias{icons$warning}
\title{\code{warning}}
\usage{
icons$warning()
}
\arguments{
\item{id}{a unique ID to be applied to the svg element}

\item{class}{a css class to be applied to the svg element}

\item{background_fill}{a color to be applied to the inner svg element}

\item{icon_color}{a color to be applied to the inner svg element}

\item{data_group}{a custom data attribute useful for subgroups of icons}

\item{aria_hidden}{a logical value for hiding elements
(Useful for purely aesthetic elements)}
}
\value{
create an svg warning icon
}
\description{
\code{warning}
}
\keyword{icons}
\keyword{warning}
