% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listbox.R
\name{listbox}
\alias{listbox}
\title{Listbox}
\usage{
listbox(
  inputId,
  title,
  label = NULL,
  options,
  values = NULL,
  classnames = NULL
)
}
\arguments{
\item{inputId}{a unique ID for the input element}

\item{title}{a title that describes the select input component}

\item{label}{a label that describes what to do (optional)}

\item{options}{an array used to generate input options}

\item{values}{an array of values to pass on to each inputs (gets \code{options}
if argument is NULL)}

\item{classnames}{a string containing one or more css classes to
attach to the returned element}
}
\value{
Create a listbox component
}
\description{
Generate a listbox widget based on user defined options.
}
\examples{
listbox(
    inputId = "popularTech",
    title = "The Most Popular Technologies",
    label = "Select a technology",
    options = c(
        "JavaScript",
        "HTML/CSS",
        "SQL",
        "Python",
        "Java",
        "Bash/Shell/Powershell",
        "C#",
        "PHP",
        "Typescript",
        "C++"
    ),
    values = c(
        "js",
        "html_css",
        "sql",
        "py",
        "java",
        "bsh_sh_powershell",
        "csharp",
        "php",
        "typescript",
        "cpp"
    ),
    classnames = "my-listbox-style"
)

}
