% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{pbar}
\alias{pbar}
\title{\code{pbar} 
R6 Class for progress bar}
\value{
R6 Class for progress bar
}
\description{
\code{pbar} 
R6 Class for progress bar

\code{pbar} 
R6 Class for progress bar
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pbar$new()}}
\item \href{#method-bar}{\code{pbar$bar()}}
\item \href{#method-listen}{\code{pbar$listen()}}
\item \href{#method-increase}{\code{pbar$increase()}}
\item \href{#method-decrease}{\code{pbar$decrease()}}
\item \href{#method-reset}{\code{pbar$reset()}}
\item \href{#method-print}{\code{pbar$print()}}
\item \href{#method-clone}{\code{pbar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$new(start = 0, min = 0, max = 7)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{the starting progress}

\item{\code{min}}{the minimum value of the progress bar}

\item{\code{max}}{the maximum value of the progress bar
\code{bar}
Create a new progress bar in the shiny UI}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bar"></a>}}
\if{latex}{\out{\hypertarget{method-bar}{}}}
\subsection{Method \code{bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$bar(
  id = NULL,
  fill = NULL,
  fixed = FALSE,
  position = "top",
  yOffset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{A unique ID for the progress bar}

\item{\code{fill}}{A color used to style the progress bar}

\item{\code{fixed}}{A logical value that is used to fix the progress to the
top or bottom of the parent element}

\item{\code{position}}{If `fixed = TRUE`, then the argument position
can be used to fix the progress bar to the "top" or "bottom" of
the parent element.}

\item{\code{yOffset}}{A CSS value used to adjust the y position of the
      progress bar relative to the parent element
\code{listen}
Initializes the progress bar within the shiny server
\code{increase}
Increase the progress bar by 1 another number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listen"></a>}}
\if{latex}{\out{\hypertarget{method-listen}{}}}
\subsection{Method \code{listen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$listen()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-increase"></a>}}
\if{latex}{\out{\hypertarget{method-increase}{}}}
\subsection{Method \code{increase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$increase(by = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by}}{A number between min and max values (default = 1)
\code{decrease}
Decrease the progress bar by 1 another number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decrease"></a>}}
\if{latex}{\out{\hypertarget{method-decrease}{}}}
\subsection{Method \code{decrease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$decrease(by = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{by}}{A number between min and max values (default = 1)
\code{reset}
resets progress bar to its initial state}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$reset(to = self$start)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{to}}{A numeric value that is used to reset progress bar
    The default is the starting value, but can be another number
    as long as it is within the min and max values of the progress
    bar.
Prints internal values of the progress bar}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pbar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
