% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressbar.R
\name{progressbar}
\alias{progressbar}
\title{Progress Bar}
\usage{
progressbar(start = 0, min = 0, max = 7)
}
\arguments{
\item{start}{the starting point for the progress bar}

\item{min}{the minimum value of the progress bar}

\item{max}{the maximum value of the progress bar}
}
\value{
Create a new progress bar
}
\description{
Create a new progress bar in your shiny app. This progress bar was designed
for apps that many pages that are viewed in a given order. This may be
useful apps that have a series of instructions screens or that are
designed for qualitative data collection (i.e., survey data). For example,
let's say that an app has 10 instruction pages. The user is instructed to
read through all pages and use the navigation buttons (next and back) to
move between pages. The progress bar can be useful in this situation as it
visually shows how many pages are left.
}
\examples{

```{r}
library(shiny)

mybar <- accessibleshiny::progressbar(max = 10)

ui <- tagList(
    accessibleshiny::use_accessibleshiny(),
    tags$main(
        style = "display: block; height: 50vh; margin-top: 50px",
        actionButton("increaseBar", "Next"),
        actionButton("descreaseBar", "Previous")
    ),
    mybar$bar(
        id = "bar1",
        fixed = TRUE
    )
)

server <- function(input, output) {
    mybar$listen()

    observeEvent(input$increaseBar, {
        mybar$increase()
    })

    observeEvent(input$descreaseBar, {
        mybar$decrease()
    })
}

shinyApp(ui, server)
```
}
