% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radios.R
\name{radios}
\alias{radios}
\title{Radio Inputs

The function \code{radios} creates a list of radio inputs, but as buttons.
This component is recommended for use over dropdown menus as
all options are visible and can be navigated with ease. This is
not recommended for use with variables with many options. There is
not limit to the number of options that can be generated, but
the number should be very low (e.g., < 7-8 ).}
\usage{
radios(
  id = "",
  css = NULL,
  inputId = c(),
  labels = c(),
  values = NULL,
  selected = NULL,
  caption = NULL
)
}
\arguments{
\item{id}{a string containing an unique id to be applied the
parent html element (i.e., fieldset; required)}

\item{css}{a string containing css class(es) to apply to the
parent html elemt (i.e., fieldset; optional; default \code{NULL})}

\item{inputId}{A string or an array containing input ids to
be passed on to each radio input.}

\item{labels}{A string or an array containing the names of each input;
this item is required.}

\item{values}{A string or an array of values to assign to each input
element (optional, but recommended).}

\item{selected}{an integer or an array of integers indicating which
input elements are selected by default (optional)}

\item{caption}{an optional string used to create a caption for the
radio group inputs (optional, but recommended.}
}
\value{
Returns html markup for a radio inputs styled as input buttons.
}
\description{
Radio Inputs

The function \code{radios} creates a list of radio inputs, but as buttons.
This component is recommended for use over dropdown menus as
all options are visible and can be navigated with ease. This is
not recommended for use with variables with many options. There is
not limit to the number of options that can be generated, but
the number should be very low (e.g., < 7-8 ).
}
\section{Arguments}{
NA
}

\examples{
x <- c("apples", "limes", "oranges", "lemons", "bananas")
y <- c("Apples", "Limes", "Oranges", "Lemons", "Bananas")
radios(
     id = "fruit",
     css = "dark-theme",
     inputId = x,
     labels = y,
     values = x,
     selected = c(1, 3, 5),
     caption = "Choose 1 or more fruits"
)

}
\author{
dcruvolo
}
\keyword{accessibleshiny,}
\keyword{inputs}
\keyword{radios,}
