% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion.R
\name{reset_accordion}
\alias{reset_accordion}
\title{\code{reset_accordion}}
\usage{
reset_accordion(inputId)
}
\arguments{
\item{inputId}{the of of the component to reset}
}
\description{
A server-side function that resets the accordion component to it's
default statue (closed).
}
