% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listbox.R
\name{reset_listbox}
\alias{reset_listbox}
\title{Reset Listbox}
\usage{
reset_listbox(inputId, value = NULL)
}
\arguments{
\item{inputId}{the ID of the listbox to reset}

\item{value}{the name of the option to select (i.e.,
\code{listbox(choices = c(...))}) or a value (i.e.,
\code{listbox(values = c(...))}). If \code{value} is left blank, then the
first option is selected.}
}
\value{
Reset a listbox to the first or named option
}
\description{
Reset the listbox component to it's initial state or a specific option.
The initial state, is the first available option.
}
