% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_accessibleshiny.R
\name{use_accessibleshiny}
\alias{use_accessibleshiny}
\title{load_dependencies}
\usage{
use_accessibleshiny()
}
\value{
Load package dependencies
}
\description{
When called, this function attaches the css files in the <head> of the html
document.
}
\examples{
use_accessiblyshiny()
}
