% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{datatable}
\alias{datatable}
\title{Accessible and Responsive Datatables}
\usage{
datatable(
  data,
  caption = NULL,
  caption_placement = NULL,
  id = NULL,
  classnames = NULL,
  row_highlighting = TRUE,
  row_headers = FALSE,
  is_responsive = TRUE,
  html_escape = TRUE
)
}
\arguments{
\item{data}{A data object used to render the table (required)}

\item{caption}{A short description (1-2 lines) for the table (optional)}

\item{caption_placement}{change the position of the caption in relation to
the table. Choices are "top" (default) or "bottom".}

\item{id}{a unique ID for the table}

\item{classnames}{a string containing one or more css classes}

\item{row_highlighting}{If \code{TRUE} (default), whenever the mouse hovers over
a cell, the entire row will be highlighted}

\item{row_headers}{If \code{TRUE}, the first cell in each table row will be
rendered as a row header (default: \code{FALSE}).}

\item{is_responsive}{If \code{TRUE} (default), the HTML structure of the table
will be responsive.}

\item{html_escape}{If \code{TRUE} (default), all cell content will be rendered
as plain text.}
}
\value{
Create a responsive datatable
}
\description{
The \code{datatable} function creates an accessible, responsive table from a
dataset. The function returns a shiny tagList object which can
be used in shiny applications or markdown documents. This function
can also be used as an html table generator and the output can be
written to file. This function takes the following arguments.
}
\examples{

```{r}
datatable(data = iris)

datatable(data = iris, id = "iris-table", classnames = "dark-theme")

datatable(data = iris, is_responsive = FALSE)

datatable(data = iris, id = "iris", row_headers = TRUE)

df <- dplyr::starwars
df$link <- paste0(
    "<a href='https://www.google.com/search?q=",
    gsub(" ", "+", df$character),
    "'>",
    df$character,
    "</a>"
)
tbl <- datatable(data = df, html_escape = FALSE)
writeLines(as.character(tbl), "~/Desktop/table.html")
```

}
