% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{icons$plus}
\alias{icons$plus}
\title{\code{plus}
A plus sign with a filled background}
\usage{
icons$plus()
}
\arguments{
\item{id}{a unique ID for the icon}

\item{class}{a css classname to assign to the icon}

\item{background_fill}{a background color for the icon (default: #3F454B)}

\item{icon_color}{a color for the plus symbol (default: #ffffff)}
}
\value{
A plus sign with a filled background
}
\description{
\code{plus}
A plus sign with a filled background
}
\keyword{icons,}
\keyword{plus}
