% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{\code{Progress}}
\value{
Create a new R6 progress bar
}
\description{
Create a new R6 progress bar.
}
\examples{
appProgress <- pbars::progress$new(min = 0, max = 12)
ui <- tagList(
     ...
     appProgress$bar(
         id = "complete",
         fixed = TRUE,
         position = "bottom",
         fill = "#006992"
     ),
     ...,
     tags$button(
         id = "nextPage",
         class = "action-button shiny-bound-input",
         "Next Page"
     )
)
server <- function(input, output, session) {
     appProgress$listen()
     observeEvent(input$nextPage, {
         appProgress$increase()
     })
}
shinyApp(ui, server)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{progress$new()}}
\item \href{#method-bar}{\code{progress$bar()}}
\item \href{#method-listen}{\code{progress$listen()}}
\item \href{#method-increase}{\code{progress$increase()}}
\item \href{#method-decrease}{\code{progress$decrease()}}
\item \href{#method-reset}{\code{progress$reset()}}
\item \href{#method-print}{\code{progress$print()}}
\item \href{#method-clone}{\code{progress$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$new(start = 0, min = 0, max = 7)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{the starting value for the progress bar}

\item{\code{min}}{the minimum value for the progress bar (default is 0)}

\item{\code{max}}{the maxium value for the progress bar (default is 7)
\code{bar}
Create a new progress bar in the shiny UI}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bar"></a>}}
\if{latex}{\out{\hypertarget{method-bar}{}}}
\subsection{Method \code{bar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$bar(
  id = NULL,
  fill = NULL,
  fixed = FALSE,
  position = "top",
  yOffset = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the progress bar}

\item{\code{fill}}{a color of the progress bar (default is #bdbdbd)}

\item{\code{fixed}}{logical value to set of the positioning of the pbar}

\item{\code{position}}{if \code{fixed = TRUE}, the bar can be placed at the
top or bottom of the window (default is top)}

\item{\code{yOffset}}{a css value to offset the y position of the progress
     bar. This is useful if you want the progress bar to display
     directly below a navigation bar or another html element
\code{listen}
Initializes Progress Bar server-side}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-listen"></a>}}
\if{latex}{\out{\hypertarget{method-listen}{}}}
\subsection{Method \code{listen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$listen(id = self$elem)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the progress bar
\code{increase}
Increase the internal counter by 1 or another number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-increase"></a>}}
\if{latex}{\out{\hypertarget{method-increase}{}}}
\subsection{Method \code{increase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$increase(id = self$elem, by = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the progress bar}

\item{\code{by}}{the number to increases by (default 1)
\code{descrease}
Decreases the internal counter by 1 or another number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decrease"></a>}}
\if{latex}{\out{\hypertarget{method-decrease}{}}}
\subsection{Method \code{decrease()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$decrease(id = self$elem, by = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the progress bar}

\item{\code{by}}{the number to decreases by (default 1)
\code{reset}
Reset a progress bar to start or another value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$reset(id = self$elem, to = self$start)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the id of the progress bar}

\item{\code{to}}{the value to reset the bar to (defaults to min value or 0)
\code{print}
Print internal values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Print internal values to the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{progress$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
